# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CleanOps Pro - Multi-tenant SaaS API for cleaning businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, complete, cancelled]
        - name: assigned_user_id
          in: query
          schema:
            type: integer
        - name: client_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Create job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update job
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /jobs/{id}/complete:
    post:
      operationId: completeJob
      tags: [jobs]
      summary: Mark job as complete (requires after photos)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Job completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
  /jobs/{id}/photos:
    get:
      operationId: getJobPhotos
      tags: [photos]
      summary: Get job photos
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Job photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoListResponse"

    post:
      operationId: uploadJobPhoto
      tags: [photos]
      summary: Upload job photo
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadPhotoRequest"
      responses:
        "201":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
...
  /loyalty/settings:
    get:
      operationId: getLoyaltySettings
      tags: [loyalty]
      summary: Get loyalty program settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Loyalty settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltySettings"

    put:
      operationId: updateLoyaltySettings
      tags: [loyalty]
      summary: Update loyalty program settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoyaltySettings"
      responses:
        "200":
          description: Loyalty settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoyaltySettings"
  /loyalty/points/{clientId}:
    get:
      operationId: getClientLoyaltyPoints
      tags: [loyalty]
      summary: Get client loyalty points and history
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client loyalty points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientLoyaltyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [owner, admin, office, technician]
        pay_rate:
          type: number
          nullable: true
        pay_type:
          type: string
          enum: [hourly, per_job, fee_split]
          nullable: true
        is_active:
          type: boolean
        hire_date:
          type: string
          format: date
          nullable: true
        avatar_url:
          type: string
          nullable: true
        productivity_pct:
          type: number
          nullable: true
      required:
        - id
        - email
        - first_name
        - last_name
        - role
        - is_active
...
    Client:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        loyalty_points:
          type: integer
        created_at:
          type: string
          format: date-time
      required:
        - id
        - first_name
        - last_name
        - loyalty_points
        - created_at
...
    Job:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_name:
          type: string
        assigned_user_id:
          type: integer
          nullable: true
        assigned_user_name:
          type: string
          nullable: true
        service_type:
          type: string
          enum: [standard_clean, deep_clean, move_out, recurring, post_construction, move_in, office_cleaning, common_areas, retail_store, medical_office, ppm_turnover, post_event]
        status:
          type: string
          enum: [scheduled, in_progress, complete, cancelled]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: string
          nullable: true
        frequency:
          type: string
          enum: [weekly, biweekly, every_3_weeks, monthly, on_demand]
        base_fee:
          type: number
        allowed_hours:
          type: number
          nullable: true
        actual_hours:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        before_photo_count:
          type: integer
        after_photo_count:
          type: integer
        created_at:
          type: string
          format: date-time
      required:
        - id
        - client_id
        - client_name
        - service_type
        - status
        - scheduled_date
        - frequency
        - base_fee
        - before_photo_count
        - after_photo_count
        - created_at
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            client:
              $ref: "#/components/schemas/Client"
            assigned_user:
              $ref: "#/components/schemas/User"
              nullable: true
            photos:
              type: array
              items:
                $ref: "#/components/schemas/Photo"
            timeclock_entries:
              type: array
              items:
                $ref: "#/components/schemas/TimeclockEntry"
            invoice:
              $ref: "#/components/schemas/Invoice"
              nullable: true
            checklist_items:
              type: array
              items:
                $ref: "#/components/schemas/ChecklistItem"
...
    JobListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateJobRequest:
      type: object
      properties:
        client_id:
          type: integer
        assigned_user_id:
          type: integer
          nullable: true
        service_type:
          type: string
          enum: [standard_clean, deep_clean, move_out, recurring, post_construction, move_in, office_cleaning, common_areas, retail_store, medical_office, ppm_turnover, post_event]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: string
          nullable: true
        frequency:
          type: string
          enum: [weekly, biweekly, every_3_weeks, monthly, on_demand]
        base_fee:
          type: number
        allowed_hours:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - client_id
        - service_type
        - scheduled_date
        - frequency
        - base_fee
    UpdateJobRequest:
      type: object
      properties:
        assigned_user_id:
          type: integer
          nullable: true
        service_type:
          type: string
          enum: [standard_clean, deep_clean, move_out, recurring, post_construction, move_in, office_cleaning, common_areas, retail_store, medical_office, ppm_turnover, post_event]
        status:
          type: string
          enum: [scheduled, in_progress, complete, cancelled]
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: string
          nullable: true
        frequency:
          type: string
          enum: [weekly, biweekly, every_3_weeks, monthly, on_demand]
        base_fee:
          type: number
        allowed_hours:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
    Photo:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: integer
        photo_type:
          type: string
          enum: [before, after]
        url:
          type: string
        timestamp:
          type: string
          format: date-time
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
      required:
        - id
        - job_id
        - photo_type
        - url
        - timestamp
    PhotoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        before_count:
          type: integer
        after_count:
          type: integer
      required:
        - data
        - before_count
        - after_count
    UploadPhotoRequest:
      type: object
      properties:
        photo_type:
          type: string
          enum: [before, after]
        data_url:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
      required:
        - photo_type
        - data_url
    TimeclockEntry:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: string
        clock_in_at:
          type: string
          format: date-time
        clock_out_at:
          type: string
          format: date-time
          nullable: true
        clock_in_lat:
          type: number
          nullable: true
        clock_in_lng:
          type: number
          nullable: true
        clock_out_lat:
          type: number
          nullable: true
        clock_out_lng:
          type: number
          nullable: true
        distance_from_job_ft:
          type: number
          nullable: true
        flagged:
          type: boolean
        duration_hours:
          type: number
          nullable: true
      required:
        - id
        - job_id
        - user_id
        - user_name
        - clock_in_at
        - flagged
...
    InvoiceLineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unit_price:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unit_price
        - total
    Invoice:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_name:
          type: string
        job_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        line_items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        tips:
          type: number
        total:
          type: number
        created_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - client_id
        - client_name
        - status
        - line_items
        - subtotal
        - tips
        - total
        - created_at
...
    LoyaltySettings:
      type: object
      properties:
        program_style:
          type: string
          enum: [points, punch_card, tiered_vip]
        pts_per_cleaning:
          type: integer
        pts_per_dollar:
          type: integer
        referral_pts:
          type: integer
        review_pts:
          type: integer
        birthday_pts:
          type: integer
        autopay_pts:
          type: integer
        enabled:
          type: boolean
      required:
        - program_style
        - pts_per_cleaning
        - pts_per_dollar
        - referral_pts
        - review_pts
        - birthday_pts
        - autopay_pts
        - enabled
    LoyaltyPointLog:
      type: object
      properties:
        id:
          type: integer
        points:
          type: integer
        action:
          type: string
          enum: [earn, redeem]
        reason:
          type: string
        job_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - points
        - action
        - reason
        - created_at
    ClientLoyaltyResponse:
      type: object
      properties:
        client_id:
          type: integer
        total_points:
          type: integer
        lifetime_points:
          type: integer
        history:
          type: array
          items:
            $ref: "#/components/schemas/LoyaltyPointLog"
      required:
        - client_id
        - total_points
        - lifetime_points
        - history
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        completed:
          type: boolean
        required:
          type: boolean
      required:
        - id
        - label
        - completed
        - required
...